made website rank a double in the code to reflect how it is stored in the database and in solr

+++ src/db.rs
@@ -25,7 +25,7 @@ pub struct Website {
     pub title: String,
     pub text: String,
     pub url: String,
-    pub rank: i32,
+    pub rank: f64,
     pub type_of_website: String
 }
